+++ service/SpaceCenter/test/test_editor.py
@@ -160,6 +160,10 @@ class TestEditorLaunchVessel(krpctest.TestCase):
         editor_vessel = self.space_center.editor.vessel
         mass = editor_vessel.mass
         moi = editor_vessel.moment_of_inertia
+        root = editor_vessel.parts.root
+        frame = root.reference_frame
+        com = editor_vessel.center_of_mass(frame)
+        part_com = root.center_of_mass(frame)
         self.space_center.editor.launch_vessel("LaunchPad")
         flight = self.space_center.active_vessel
         self.assertEqual(craft, flight.name)
@@ -171,6 +175,13 @@ class TestEditorLaunchVessel(krpctest.TestCase):
                 flight_moi[i],
                 delta=max(abs(moi[i]) * 0.05, 1),
             )
+        flight_root = flight.parts.root
+        flight_frame = flight_root.reference_frame
+        flight_com = flight.center_of_mass(flight_frame)
+        flight_part_com = flight_root.center_of_mass(flight_frame)
+        for i in range(3):
+            self.assertAlmostEqual(com[i], flight_com[i], delta=0.5)
+            self.assertAlmostEqual(part_com[i], flight_part_com[i], delta=0.2)
 
 
 if __name__ == "__main__":